cleanup of scripts and data
