+++ src/design/__init__.py
@@ -1,5 +1,6 @@
 # pyright: reportUnusedImport=false
-from design.data import SCRIPTS, Script
+from design.Script import Script
+from design.data import SCRIPTS, SORTING_ORDER
 from design.Item import Item
 from design.Job import Job
 from design.Test import Test, ScriptError
